employee_management_system_source/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.20
Werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.0.0
Flask-Migrate==4.0.4
PyJWT==2.8.0
